refactor: reorganize code-highlight tests and update tree-sitter-diff dependency
- Move bash syntax highlighting tests from highlight.rs to dedicated tests/bash.rs module
- Extract diff parsing tests from diff_test.rs to tests/diff.rs with enhanced assertions
- Remove monolithic highlight.rs test module in favor of modular test organization
- Update tree-sitter-diff dependency to use git source with v0.1.0 tag for better maintenance
- Add Cargo patch section for tree-sitter-diff to ensure consistent dependency resolution
- Delete obsolete diff_test.rs file after test migration
- Update lib.rs to reference new tests module structure

This reorganization improves test maintainability by separating syntax highlighting
tests by language and provides clearer test organization for future language support.
