+++ Cargo.toml
@@ -67,3 +67,6 @@ similar = "2.7.0"
 [dev-dependencies]
 ctor.workspace = true
 tempfile = "3.22.0"
+
+[patch.crates-io]
+tree-sitter-diff = { git = "https://github.com/tree-sitter-grammars/tree-sitter-diff.git", tag = "v0.1.0" }
