+++ crates/code-highlight/src/tests.rs
@@ -1,0 +1,2 @@
+mod bash;
+mod diff;
